Added a setup_build script for testing the various GoClone projects.
Support is provided for testing projects with mysql or mysql2 (adapter)
or sqlite3. Gem testing is also provided (i.e. don't bother setting up
a database.yml).

Database YML creation is based on generating random user names and
databases. These are but based on the project name, hence each project
gets the same random database user between builds.

Highlight the output of the setup_build script so that we can differentiate
it and the testing output.

Automagically any '*.ymls.sample' files are linked to their correct yml
names but this can be deactivated.

On branch: master