Added a deleted_at field to the projets to indicate that a project is about to be deleted (projects are DELETE FROM in the bigtuna world, so the deleted_at field is not to be considered as acts_as_paranoid).
When a project is destroyed, the field is set and a delayed job is setup
to delete the project. In the project overview the project is not shown
since it's deleted at field is set (scope not_pending_deletion on Project).

This was needed because project deleteion on the server was taking ages and
blocking the browser request, even leading to request timeouts.

On branch: master